+++ Classes/TiGamekitModule.m
@@ -17,19 +17,6 @@ gameCenterAuthenticationComplete;
 
 #pragma mark Internal
 
-BOOL isGameCenterAPIAvailable()
-{
-    // Check for presence of GKLocalPlayer API.
-    Class gcClass = (NSClassFromString(@"GKLocalPlayer"));
-    
-    // The device must be running running iOS 4.1 or later.
-    NSString *reqSysVer = @"4.1";
-    NSString *currSysVer = [[UIDevice currentDevice] systemVersion];
-    BOOL osVersionSupported = ([currSysVer compare:reqSysVer options:NSNumericSearch] != NSOrderedAscending);
-    
-    return (gcClass && osVersionSupported); 
-}
-
 // this is generated for your module, please do not change it
 -(id)moduleGUID
 {
@@ -46,21 +33,8 @@ BOOL isGameCenterAPIAvailable()
 
 -(void)startup
 {
-	// this method is called when the module is first loaded
-	// you *must* call the superclass
 	[super startup];
 	[self isPlayerAuthenticated ];
-	NSLog(@"[INFO] %@ loaded",self);
-}
-
--(void)shutdown:(id)sender
-{
-	// this method is called when the module is being unloaded
-	// typically this is during shutdown. make sure you don't do too
-	// much processing here or the app will be quit forceably
-	
-	// you *must* call the superclass
-	[super shutdown:sender];
 }
 
 #pragma mark Cleanup 
@@ -73,35 +47,20 @@ BOOL isGameCenterAPIAvailable()
 	[super dealloc];
 }
 
-#pragma mark Internal Memory Management
-
--(void)didReceiveMemoryWarning:(NSNotification*)notification
-{
-	// optionally release any resources that can be dynamically
-	// reloaded once memory is available - such as caches
-	[super didReceiveMemoryWarning:notification];
-}
-
 #pragma mark Listener Notifications
 
-/*-(void)_listenerAdded:(NSString *)type count:(int)count
+-(id)isAvailable:(id)args
 {
-	if (count == 1 && [type isEqualToString:@"my_event"])
-	{
-		// the first (of potentially many) listener is being added 
-		// for event named 'my_event'
-	}
+    // Check for presence of GKLocalPlayer API.
+    Class gcClass = (NSClassFromString(@"GKLocalPlayer"));
+    
+    // The device must be running running iOS 4.1 or later.
+    NSString *reqSysVer = @"4.1";
+    NSString *currSysVer = [[UIDevice currentDevice] systemVersion];
+    BOOL osVersionSupported = ([currSysVer compare:reqSysVer options:NSNumericSearch] != NSOrderedAscending);
+    
+    return NUMBOOL(gcClass && osVersionSupported); 
 }
-
--(void)_listenerRemoved:(NSString *)type count:(int)count
-{
-	if (count == 0 && [type isEqualToString:@"my_event"])
-	{
-		// the last listener called for event named 'my_event' has
-		// been removed, we can optionally clean up any resources
-		// since no body is listening at this point for that event
-	}
-} */
 
 - (BOOL)isPlayerAuthenticated
 {
@@ -140,65 +99,19 @@ BOOL isGameCenterAPIAvailable()
                     self.currentPlayerID = localPlayer.playerID;
                     
                     // Load game instance for new current player, if none exists create a new.
-					[self fireEvent:@"loggedin"];
+					[self fireEvent:@"loggedIn"];
                 }
             } else {     
                 // No user is logged into Game Center, run without Game Center support or user interface. 
                 self.gameCenterAuthenticationComplete = NO;
+                
+                [self fireEvent:@"notLoggedIn"];
             }
         }];
     //}    
     
     // The user is not authenticated until the Completion Handler block is called. 
     return YES;
-}
-
-#pragma Public APIs
-
-- (void) achievementSubmitted: (GKAchievement*) ach error:(NSError*) error;
-{
-	if((error == NULL) && (ach != NULL))
-	{
-		if(ach.percentComplete == 100.0)
-		{
-			/*[self showAlertWithTitle: @"Achievement Earned!"
-							 message: [NSString stringWithFormat: @"Great job!  You earned an achievement: \"%@\"", NSLocalizedString(ach.identifier, NULL)]];*/
-			[self fireEvent:@"achievementEarned"];
-		}
-		else
-		{
-			if(ach.percentComplete > 0)
-			{
-				/*[self showAlertWithTitle: @"Achievement Progress!"
-								 message: [NSString stringWithFormat: @"Great job!  You're %.0f\%% of the way to: \"%@\"",ach.percentComplete, NSLocalizedString(ach.identifier, NULL)]];*/
-				[self fireEvent:@"achievementProgress"];
-			}
-		}
-	}
-	else
-	{
-		/*[self showAlertWithTitle: @"Achievement Submission Failed!"
-						 message: [NSString stringWithFormat: @"Reason: %@", [error localizedDescription]]];*/
-		[self fireEvent:@"achievementFailed"];
-	}
-}
-
-
--(id)example:(id)args
-{
-	// example method
-	return @"hello world";
-}
-
--(id)exampleProp
-{
-	// example property getter
-	return @"hello world";
-}
-
--(void)exampleProp:(id)value
-{
-	// example property setter
 }
 
 @end