+++ documentation/changelog.md
@@ -1,3 +1,10 @@
 # Change Log
 <pre>
+v1.1	Updated documentation for greater clarity and consistency with other Titanium modules.
+		Fixed achievements. [MOD-327]
+		Added the "isAvailable()" method to the module.
+		BREAKING CHANGE: Renamed the "loggedin" event to "loggedIn" (with an upper case I) for platform consistency.
+		Added the new "notLoggedIn" event.
+		BREAKING CHANGE: Renamed "dataRecieved" event to "dataReceived" (spelling fix).
+		
 v1.0    Initial Release